# codeaware-py/openapi/current-release.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryHit": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "memory_type": {
            "$ref": "#/components/schemas/MemoryType"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "memory_type",
          "source",
          "similarity"
        ],
        "title": "MemoryHit",
        "type": "object"
      },
      "MemorySaveRequest": {
        "properties": {
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "memory_type": {
            "const": "REFERENCE",
            "default": "REFERENCE",
            "title": "Memory Type",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "required": [
          "content"
        ],
        "title": "MemorySaveRequest",
        "type": "object"
      },
      "MemorySaveVO": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "content"
        ],
        "title": "MemorySaveVO",
        "type": "object"
      },
      "MemoryType": {
        "description": "长期记忆来源类型（ADR-0001）。",
        "enum": [
          "REFERENCE",
          "FACT"
        ],
        "title": "MemoryType",
        "type": "string"
      },
...
      "PromptCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "review_dimensions": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Dimensions"
          },
          "role_setting": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Role Setting",
            "type": "string"
          },
          "severity_levels": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Levels"
          },
          "template_body": {
            "maxLength": 60000,
            "minLength": 1,
            "title": "Template Body",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PromptType"
          }
        },
        "required": [
          "type",
          "name",
          "role_setting",
          "template_body"
        ],
        "title": "PromptCreateRequest",
        "type": "object"
      },
      "PromptPreviewVO": {
        "properties": {
          "rendered": {
            "title": "Rendered",
            "type": "string"
          }
        },
        "required": [
          "rendered"
        ],
        "title": "PromptPreviewVO",
        "type": "object"
      },
      "PromptTemplateVO": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "review_dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Dimensions"
          },
          "role_setting": {
            "title": "Role Setting",
            "type": "string"
          },
          "severity_levels": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Levels"
          },
          "template_body": {
            "title": "Template Body",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PromptType"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "version",
          "name",
          "role_setting",
          "template_body",
          "is_active",
          "created_at"
        ],
        "title": "PromptTemplateVO",
        "type": "object"
      },
      "PromptType": {
        "description": "Prompt 模板类型 = 逻辑身份（ADR-0005）。",
        "enum": [
          "CODE_REVIEW",
          "UNIT_TEST",
          "AI_README",
          "CHAT"
        ],
        "title": "PromptType",
        "type": "string"
      },
...
      "Result_MemorySaveVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemorySaveVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[MemorySaveVO]",
        "type": "object"
      },
      "Result_NoneType_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[NoneType]",
        "type": "object"
      },
...
      "Result_PromptPreviewVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptPreviewVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[PromptPreviewVO]",
        "type": "object"
      },
      "Result_PromptTemplateVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptTemplateVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[PromptTemplateVO]",
        "type": "object"
      },
...
      "Result_list_MemoryHit__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MemoryHit"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[list[MemoryHit]]",
        "type": "object"
      },
      "Result_list_PromptTemplateVO__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PromptTemplateVO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[list[PromptTemplateVO]]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CodeAware - AI 驱动的研发效能平台 (Python 重构)",
    "title": "codeaware",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/memory/long-term": {
      "post": {
        "operationId": "save_long_term_api_memory_long_term_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_MemorySaveVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Long Term",
        "tags": [
          "Memory"
        ]
      }
    },
    "/api/memory/long-term/search": {
      "get": {
        "operationId": "search_long_term_api_memory_long_term_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "threshold",
            "required": false,
            "schema": {
              "default": 0.3,
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Threshold",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Top K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_MemoryHit__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Long Term",
        "tags": [
          "Memory"
        ]
      }
    },
    "/api/memory/long-term/{memory_id}": {
      "delete": {
        "operationId": "delete_long_term_api_memory_long_term__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Long Term",
        "tags": [
          "Memory"
        ]
      }
    },
    "/api/prompts": {
      "get": {
        "operationId": "list_prompts_api_prompts_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PromptType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_PromptTemplateVO__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Prompts",
        "tags": [
          "Prompt"
        ]
      },
      "post": {
        "operationId": "create_prompt_api_prompts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PromptTemplateVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Prompt",
        "tags": [
          "Prompt"
        ]
      }
    },
    "/api/prompts/{template_id}/activate": {
      "post": {
        "operationId": "activate_api_prompts__template_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PromptTemplateVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activate",
        "tags": [
          "Prompt"
        ]
      }
    },
    "/api/prompts/{template_id}/preview": {
      "get": {
        "operationId": "preview_api_prompts__template_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sample_code",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 20000,
              "title": "Sample Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PromptPreviewVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview",
        "tags": [
          "Prompt"
        ]
      }
    },
...
  }
}